@@ -10,15 +10,17 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 class HoneypotValidationListener implements EventSubscriberInterface
 {
-    private $translator;
-    private $translationDomain;
-    private $fieldName;
-    private $errorMessage;
+    private ?TranslatorInterface $translator;
+    private string $translationDomain;
+    private string $fieldName;
+    private string $errorMessage;
 
-    public function __construct(TranslatorInterface $translator = null,
-                                $translationDomain,
-                                $fieldName,
-                                $errorMessage)
+    public function __construct(
+        ?TranslatorInterface $translator,
+        string $translationDomain,
+        string $fieldName,
+        string $errorMessage
+    )
     {
         $this->translator = $translator;
         $this->translationDomain = $translationDomain;
@@ -26,7 +28,7 @@ class HoneypotValidationListener implements EventSubscriberInterface
         $this->errorMessage = $errorMessage;
     }
 
-    public function preSubmit(FormEvent $event)
+    public function preSubmit(FormEvent $event): void
     {
         $form = $event->getForm();
         $data = $event->getData();
@@ -50,7 +52,7 @@ class HoneypotValidationListener implements EventSubscriberInterface
         $event->setData($data);
     }
 
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return array(
             FormEvents::PRE_SUBMIT => 'preSubmit',
