@@ -2,33 +2,39 @@
 
 namespace Isometriks\Bundle\SpamBundle\Form\Extension\Spam\Provider;
 
+use DateTime;
+
 interface TimedSpamProviderInterface
 {
     /**
-     * @return \DateTime $startTime
+     * Generate form time.
+     *
+     * @param string $name
+     *
+     * @return DateTime
      */
-    public function generateFormTime($name);
+    public function generateFormTime(string $name): DateTime;
 
     /**
      * Check if form has time.
      *
      * @param string $name
      */
-    public function hasFormTime($name);
+    public function hasFormTime(string $name): bool;
 
     /**
      * Gets the form time for specified form.
      *
-     * @param $name Name of form to get
+     * @param string $name Name of form to get
      */
-    public function getFormTime($name);
+    public function getFormTime(string $name);
 
     /**
      * Removes a form time name.
      *
      * @param string $name
      */
-    public function removeFormTime($name);
+    public function removeFormTime(string $name): void;
 
     /**
      * Check if form time is valid.
@@ -38,5 +44,5 @@ interface TimedSpamProviderInterface
      *
      * @return bool $valid
      */
-    public function isFormTimeValid($name, array $options);
+    public function isFormTimeValid(string $name, array $options): bool;
 }
